+++ R/groups.R
@@ -1,3 +1,0 @@
-#' @import purrr vctrs dplyr
-#' @importFrom rlang syms list2 abort
-NULL