+++ tests/testthat/test_igroup.R
@@ -6,7 +6,7 @@ tsti_a <- tibble::tribble(
   "A", "y", 3,
   "B", NA, 3
 ) %>%
-  group_by2(main = NULL, grp = NA)
+  group_by2(main, grp = NA)
 
 tsti_b <-
   tibble::tribble(
@@ -17,20 +17,24 @@ tsti_b <-
     2, "sub", "sub2", 3.8,
     2, "main", "na", 4.0
   ) %>%
-  group_by2(grp = NULL, subgrp = "na")
+  group_by2(grp, subgrp = "na")
 
 test_that("group_by2 throws if no data", {
-  expect_error(group_by2(), class="error_bad_argument")
+  expect_error(group_by2())
 })
 
 test_that("group_by2 throws if groups are unavailable", {
-  expect_error(group_by2(mtcars, nonexistent = NULL),
+  expect_error(group_by2(mtcars, nonexistent),
                class = "error_miss_col")
 })
 
+test_that("group_by2 throws if inapplicable value doesn't exist", {
+  expect_error(group_by2(mtcars, vs = 3))
+})
+
 test_that("group_by2 can make multiple groups", {
-  res <- group_by2(mtcars, vs = NULL, gear = NULL)
-  expect_equal(length(groups(res)), 2)
+  res <- group_by2(mtcars, vs, gear)
+  expect_equal(length(dplyr::groups(res)), 2)
 })
 
 test_that("can ungroup", {
@@ -45,8 +49,8 @@ test_that("group_by2 ungroups if empty args", {
 })
 
 test_that("can expand an igrouped dataframe", {
-  res <- select(ungroup(expand_igrps(tsti_b)), subgrp, val)
-  cmp <- tibble::tibble(
+  res <- dplyr::select(ungroup(expand_igrps(tsti_b)), subgrp, val)
+  cmp <- tibble(
     subgrp = c("sub1", "sub2", "sub1", "sub2", "sub1", "sub2"),
     val = c(1.9, 3.1, 9.7, 3.8, 4.0, 4.0)
   )
@@ -58,4 +62,10 @@ test_that("can expand a different igroup (?)", {
     expand_igrps(group_by2(mtcars, vs = 1))
     TRUE
   })
+})
+
+test_that("can print a grouped tbl", {
+  expect_equal(
+    grep("Row indices", paste0(format(tsti_a), collapse = "")), 1
+  )
 })