+++ NEWS.md
@@ -1,4 +1,4 @@
-# groups 0.1.0 (in development)
+# groupr 0.1.0 (in development)
 
 * S3 class `igrouped_df` which describes a grouping that can be inapplicable
 
@@ -7,5 +7,7 @@
 * S3 methods `ungroup()`, `format()`, `group_data()` for `igrouped_df`
 
 * Function `pivot_grps()` for pivoting data using a group framework
+
+* Functions `colgrp()`, `infer_colgrps()`, and `sep_colgrp()` for column grouping
 
 * `introduction` vignette